New ServUO 29.04.2020

+++ Scripts/Items/Addons/Aquarium/Fish/AlbinoCourtesanFish.cs
@@ -1,9 +1,7 @@
-using System;
-
 namespace Server.Items
 {
     public class AlbinoCourtesanFish : BaseFish
-    { 
+    {
         [Constructable]
         public AlbinoCourtesanFish()
             : base(0x3B04)
@@ -12,15 +10,15 @@ namespace Server.Items
 
         public AlbinoCourtesanFish(Serial serial)
             : base(serial)
-        { 
+        {
         }
 
         public override int LabelNumber => 1074592; // Albino Courtesan Fish
-		
+
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
