+++ Scripts/Items/Addons/Aquarium/Fish/AlbinoFrog.cs
@@ -1,9 +1,7 @@
-using System;
-
 namespace Server.Items
 {
     public class AlbinoFrog : BaseFish
-    { 
+    {
         [Constructable]
         public AlbinoFrog()
             : base(0x3B0D)
@@ -13,15 +11,15 @@ namespace Server.Items
 
         public AlbinoFrog(Serial serial)
             : base(serial)
-        { 
+        {
         }
 
         public override int LabelNumber => 1073824; // An Albino Frog
-		
+
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
