+++ Scripts/Items/Decorative/SnowyTree.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace Server.Items
 {
     public class SnowyTree : Item
@@ -27,7 +25,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
