+++ Scripts/Items/Equipment/Spellbooks/BookOfChivalry.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace Server.Items
 {
     public class BookOfChivalry : Spellbook
@@ -28,7 +26,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)1); // version
+            writer.Write(1); // version
         }
 
         public override void Deserialize(GenericReader reader)
