+++ Scripts/Items/Equipment/Weapons/BaseKnife.cs
@@ -1,4 +1,3 @@
-using System;
 using Server.Targets;
 
 namespace Server.Items
@@ -27,7 +26,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
