+++ Scripts/Items/Equipment/Weapons/BaseSword.cs
@@ -1,4 +1,3 @@
-using System;
 using Server.Targets;
 
 namespace Server.Items
@@ -24,7 +23,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
