+++ Scripts/Items/Internal/BlightedGroveTeleporters.cs
@@ -1,9 +1,7 @@
-using System;
-
 namespace Server.Items
-{ 
+{
     public class BlightedGroveTele : Teleporter
-    { 
+    {
         [Constructable]
         public BlightedGroveTele()
             : base()
@@ -17,15 +15,15 @@ namespace Server.Items
 
         public static BoneMachete GetBoneMachete(Mobile m)
         {
-            for (int i = 0; i < m.Items.Count; i ++)
+            for (int i = 0; i < m.Items.Count; i++)
             {
                 if (m.Items[i] is BoneMachete)
                     return (BoneMachete)m.Items[i];
             }
-			
+
             if (m.Backpack != null)
                 return m.Backpack.FindItemByType(typeof(BoneMachete), true) as BoneMachete;
-				
+
             return null;
         }
 
@@ -36,9 +34,9 @@ namespace Server.Items
                 m.SendLocalizedMessage(1042753, "Blighted Grove"); // ~1_SOMETHING~ has been temporarily disabled.
                 return true;
             }
-			
+
             BoneMachete machete = GetBoneMachete(m);
-			
+
             if (machete != null)
             {
                 if (0.6 > Utility.RandomDouble())
@@ -50,12 +48,12 @@ namespace Server.Items
                     machete.Delete();
                     m.SendLocalizedMessage(1075007); // Your bone handled machete snaps in half as you force your way through the poisonous undergrowth.
                 }
-				
+
                 return base.OnMoveOver(m);
             }
             else
                 m.SendLocalizedMessage(1074275); // You are unable to push your way through the tangling roots of the mighty tree.
-				
+
             return true;
         }
 
@@ -63,19 +61,19 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
         {
             base.Deserialize(reader);
-			
+
             int version = reader.ReadInt();
         }
     }
 
     public class BlightedGroveTreeInTele : Teleporter
-    { 
+    {
         [Constructable]
         public BlightedGroveTreeInTele()
             : base()
@@ -97,19 +95,19 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
         {
             base.Deserialize(reader);
-			
+
             int version = reader.ReadInt();
         }
     }
 
     public class BlightedGroveTreeOutTele : Teleporter
-    { 
+    {
         [Constructable]
         public BlightedGroveTreeOutTele()
             : base()
@@ -131,13 +129,13 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
         {
             base.Deserialize(reader);
-			
+
             int version = reader.ReadInt();
         }
     }
