+++ Scripts/Items/Internal/PaintedCavesTeleporters.cs
@@ -1,9 +1,7 @@
-using System;
-
 namespace Server.Items
 {
     public class PaintedCavesTele : Teleporter
-    { 
+    {
         [Constructable]
         public PaintedCavesTele()
             : base(new Point3D(6308, 892, 0), Map.Trammel)
@@ -22,7 +20,7 @@ namespace Server.Items
                 m.SendLocalizedMessage(1042753, "Painted Caves"); // ~1_SOMETHING~ has been temporarily disabled.
                 return true;
             }
-			
+
             return base.OnMoveOver(m);
         }
 
@@ -30,13 +28,13 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
         {
             base.Deserialize(reader);
-			
+
             int version = reader.ReadInt();
         }
     }
