+++ Scripts/Mobiles/Named/Tangle.cs
@@ -1,4 +1,3 @@
-using System;
 using Server.Items;
 
 namespace Server.Mobiles
@@ -48,7 +47,7 @@ namespace Server.Mobiles
         {
         }
 
-        public override bool GivesMLMinorArtifact => true; 
+        public override bool GivesMLMinorArtifact => true;
 
         public override Poison PoisonImmune => Poison.Lethal;
 
@@ -68,7 +67,7 @@ namespace Server.Mobiles
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
