+++ Scripts/Mobiles/Normal/AcidSlug.cs
@@ -1,4 +1,3 @@
-using System;
 using Server.Items;
 
 namespace Server.Mobiles
@@ -21,13 +20,13 @@ namespace Server.Mobiles
                 case 3: Hue = 245; break;
             }
 
-			SetStr(213, 294);
-			SetDex(80, 82);
+            SetStr(213, 294);
+            SetDex(80, 82);
             SetInt(18, 22);
 
-			SetHits(333, 370);
+            SetHits(333, 370);
 
-			SetDamage(21, 28);
+            SetDamage(21, 28);
 
             SetDamageType(ResistanceType.Physical, 100);
 
@@ -91,7 +90,7 @@ namespace Server.Mobiles
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
