+++ Scripts/Mobiles/Void Creatures/Anlorvaglem.cs
@@ -1,4 +1,3 @@
-using System;
 using Server.Items;
 
 namespace Server.Mobiles
@@ -6,8 +5,8 @@ namespace Server.Mobiles
     [CorpseName("an anlorvaglem corpse")]
     public class Anlorvaglem : BaseVoidCreature
     {
-        public override VoidEvolution Evolution => VoidEvolution.Grouping; 
-        public override int Stage => 3; 
+        public override VoidEvolution Evolution => VoidEvolution.Grouping;
+        public override int Stage => 3;
 
         [Constructable]
         public Anlorvaglem()
@@ -68,7 +67,7 @@ namespace Server.Mobiles
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
