+++ Scripts/Mobiles/Void Creatures/Anlorzen.cs
@@ -1,4 +1,3 @@
-using System;
 using Server.Items;
 
 namespace Server.Mobiles
@@ -6,8 +5,8 @@ namespace Server.Mobiles
     [CorpseName("an anlorzen corpse")]
     public class Anlorzen : BaseVoidCreature
     {
-        public override VoidEvolution Evolution => VoidEvolution.Grouping; 
-        public override int Stage => 1; 
+        public override VoidEvolution Evolution => VoidEvolution.Grouping;
+        public override int Stage => 1;
 
         [Constructable]
         public Anlorzen()
@@ -66,7 +65,7 @@ namespace Server.Mobiles
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
