+++ Scripts/Mobiles/Void Creatures/Betballem.cs
@@ -1,4 +1,3 @@
-using System;
 using Server.Items;
 
 namespace Server.Mobiles
@@ -6,8 +5,8 @@ namespace Server.Mobiles
     [CorpseName("a betballem corpse")]
     public class Betballem : BaseVoidCreature
     {
-        public override VoidEvolution Evolution => VoidEvolution.Killing; 
-        public override int Stage => 1; 
+        public override VoidEvolution Evolution => VoidEvolution.Killing;
+        public override int Stage => 1;
 
         [Constructable]
         public Betballem()
@@ -109,7 +108,7 @@ namespace Server.Mobiles
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
