+++ Scripts/Mobiles/Void Creatures/Korpre.cs
@@ -1,14 +1,10 @@
-using System;
-using Server.Items;
-using System.Collections.Generic;
-
 namespace Server.Mobiles
 {
     [CorpseName("a korpre corpse")]
     public class Korpre : BaseVoidCreature
     {
-        public override VoidEvolution Evolution => VoidEvolution.None; 
-        public override int Stage => 0; 
+        public override VoidEvolution Evolution => VoidEvolution.None;
+        public override int Stage => 0;
 
         [Constructable]
         public Korpre()
@@ -61,7 +57,7 @@ namespace Server.Mobiles
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
