+++ Scripts/Services/Quests/QuestSerializer.cs
@@ -20,7 +20,7 @@ namespace Server.Engines.Quests
         {
             if (type == null)
             {
-                writer.WriteEncodedInt((int)0x00);
+                writer.WriteEncodedInt(0x00);
             }
             else
             {
@@ -28,13 +28,13 @@ namespace Server.Engines.Quests
                 {
                     if (referenceTable[i] == type)
                     {
-                        writer.WriteEncodedInt((int)0x01);
-                        writer.WriteEncodedInt((int)i);
+                        writer.WriteEncodedInt(0x01);
+                        writer.WriteEncodedInt(i);
                         return;
                     }
                 }
 
-                writer.WriteEncodedInt((int)0x02);
+                writer.WriteEncodedInt(0x02);
                 writer.Write(type.FullName);
             }
         }
@@ -43,7 +43,7 @@ namespace Server.Engines.Quests
         {
             int encoding = reader.ReadEncodedInt();
 
-            switch ( encoding )
+            switch (encoding)
             {
                 default:
                 case 0x00: // null
@@ -75,7 +75,7 @@ namespace Server.Engines.Quests
         {
             int encoding = reader.ReadEncodedInt();
 
-            switch ( encoding )
+            switch (encoding)
             {
                 default:
                 case 0x00: // null
@@ -116,7 +116,7 @@ namespace Server.Engines.Quests
         {
             int encoding = reader.ReadEncodedInt();
 
-            switch ( encoding )
+            switch (encoding)
             {
                 default:
                 case 0x00: // null
@@ -157,7 +157,7 @@ namespace Server.Engines.Quests
         {
             int encoding = reader.ReadEncodedInt();
 
-            switch ( encoding )
+            switch (encoding)
             {
                 default:
                 case 0x00: // null
