+++ Scripts/Services/Revamped Dungeons/TheExodusEncounter/Mobiles/ExodusDrone.cs
@@ -1,4 +1,3 @@
-using System;
 using Server.Items;
 
 namespace Server.Mobiles
@@ -63,17 +62,17 @@ namespace Server.Mobiles
         {
         }
 
-        public bool FieldActive => m_FieldActive; 
-		
+        public bool FieldActive => m_FieldActive;
+
         public bool CanUseField => Hits >= HitsMax * 9 / 10; // TODO: an OSI bug prevents to verify this
-		
-        public override bool IsScaredOfScaryThings => false; 
-		
-        public override bool IsScaryToPets => true; 
-		
-        public override bool BardImmune => false; 
-		
-        public override Poison PoisonImmune => Poison.Lethal; 
+
+        public override bool IsScaredOfScaryThings => false;
+
+        public override bool IsScaryToPets => true;
+
+        public override bool BardImmune => false;
+
+        public override Poison PoisonImmune => Poison.Lethal;
 
         public override int GetIdleSound() { return 0x218; }
         public override int GetAngerSound() { return 0x26C; }
@@ -163,7 +162,7 @@ namespace Server.Mobiles
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
